# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "cwa-dcc-server",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/version/v1/test/{testId}/dcc": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Upload a DCC for a test.",
        "description": "Endpoint to upload components to build the DCC.",
        "operationId": "uploadDcc",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "description": "ID of the test (hashed GUID).",
            "required": true,
            "schema": {
              "pattern": "^[XxA-Fa-f0-9]([A-Fa-f0-9]{63})$",
              "type": "string"
            }
          },
          {
            "name": "X-CWA-PARTNER-ID",
            "in": "header",
            "description": "PartnerID. This needs only to be set if DCC-Server is contacted without DCC-Proxy in between.",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9]{1,64}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DccUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid Data format"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "415": {
            "description": "Unsupported Media Type"
          },
          "405": {
            "description": "Method Not Allowed"
          },
          "201": {
            "description": "DCC created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DccUploadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Test does not exists",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DccUploadResponse"
                }
              }
            }
          },
          "409": {
            "description": "DCC already exists",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DccUploadResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DccUploadRequest": {
        "required": [
          "dataEncryptionKey",
          "encryptedDcc"
        ],
        "type": "object",
        "properties": {
          "dccHash": {
            "pattern": "^[A-Fa-f0-9]{64}$",
            "type": "string",
            "description": "SHA256 Hash of plain DCC payload."
          },
          "encryptedDcc": {
            "maxLength": 1000,
            "minLength": 0,
            "pattern": "^[A-Za-z0-9+/=]*$",
            "type": "string",
            "description": "Base64 encoded encrypted DCC payload."
          },
          "dataEncryptionKey": {
            "maxLength": 255,
            "minLength": 0,
            "pattern": "^[A-Za-z0-9+/=]*$",
            "type": "string",
            "description": "Base64 encoded with PublicKey encrypted Data Encryption Key for encrypted DCC."
          }
        },
        "description": "Request payload to upload a DCC from laboratory."
      },
      "DccUploadResponse": {
        "type": "object",
        "properties": {
          "partialDcc": {
            "type": "string",
            "description": "Base64 encoded DCC Structure without the payload. (COSE/CBOR, Payload needs to be replaced by laboratory)"
          }
        },
        "description": "Response for uploaded DCC data."
      },
...
    }
  }
}